(#46) add a connect+wait header time out
When the windows firewall is on, and winrm is not blocked, the connection
will be accepted but no response will be sent by the winrm server.
In this case, the winrm client will wait almost indefinitely.
This prevents packer from being able to detect when winrm is ready on
a prepared VM.
The default timeout is 60s (which should be enough for running a command
on a winrm enabled host), but still not too much so that Packer doesn't
wait too long.
The timeout is overridable (check -timeout for the winrm command, or through
the Endpoint mechanism with NewEndpointWithTimeout).