Fix #48 do not abort when running long commands
Long commands that don't output any data in Parameters.Timeout time
will get an OperationTimeout error from the WinRM server.
We used to support this, but this got broken somehow.
This changeset allows to gracefully handle this case, with tests so that
we won't break this feature again.
This changeset also brings returning an error and exit code if the
connection to the WinRM server is dropped (for instance when restarting
the host).