Fix #49 split the library and the command-line tool
This is a breaking change. The tool has been moved to the winrm-cli
project (https://github.com/masterzen/winrm-cli).
This commit removes the tool, and moves back all the library code
to the github.com/masterzen/winrm package.

Dependent libraries or tool might need to be modified to take the new
package path into account.