Allow override of TLS certificate hostname
When verifying a certificate that has been passed directly with the
CACert field it is useful to be able to override the hostname without
disabling any other certificate verification checks.

For example: EC2 instances that are given a certificate during
provisioning when the IP is not yet known and there will be no DNS
entry.