Fix #77 - fix RunWithString race condition
This doesn't happen with Go < 1.10, but there was a race condition in
`RunWithString` and `RunWithInput` which had already been fixed in `Run`.
It was possible for the function to return while the `stdout` or `stderr`
copy was still in progress. Since `bytes.Buffer` is not thread safe, the
`outwriter.String()` could see an intermediate buffer state while it
was growing, leading to strange results.