More idiomatic commandWriter
To adhere a bit more to the Golang io.Writer and io.Reader interface
I changed WriteClose to be separated again into Write and Close.

I also noticed that the commandWriter would not support multiple
writers, so this is now protected by a mutex.

Still to be more idiomatic, Write will now error if less data has
been written then provided or when writing to a closed pipe.

Also, since command.Close was bugged, commands were never closed in
the Run function. With a properly working Close, the 3 functions
now correctly call Close.